Initialize the MSYS2 Git repository
The idea of this repository is that it is easy to clone via Git for
build agents that are orchestrated by some Continuous Integration
platform. Much easier, for sure, than to run an interactive MSYS2
installer, and to keep MSYS2 up-to-date via `pacman -Syu` (which may
require the user to quit the MSYS2 Bash window the hard way, via the "X"
in the window's title bar).

So what we will do is try to keep this repository synchronized with the
state of the Git for Windows SDK that we will in turn try to keep up
with MSYS2 via `pacman -Syu` (with a couple of tricks to avoid any
interactive operations).

To initialize this repository, let's first make sure that none of the
files in /usr/src/ are committed: these include the Git sources, Git for
Windows' build-extra sources, and of course the MINGW-packages and
MSYS2-packages sources of MSYS2.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>