Add Pacman's local GPG files
These files are typically generated when initializing Pacman's keys. The
idea of this repository is to side-step all of the interactive tasks
required to initialize a Git for Windows SDK, so let's just commit these
here files.

Note: we must never, *ever* use this here SDK for anything else than for
build agents, and certainly we must make sure that no other secret keys
are added.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>