Commit edits by git-extra-1.1.118.114b5d3-1's post-install script
The way Git for Windows' SDK works is that it adapts an MSYS2
installation to its needs. Unfortunately, not all of these settings are
possible without editing files owned by certain MSYS2 packages.

For example, /etc/nsswitch.conf is owned by the filesystem package, yet
we really need to modify it because the version provided by the
filesystem package is just not good enough for Git for Windows'
purposes.

Modifications like these are hence required, and performed via
git-extra's post-install script (which, as a consequence, should always
be reinstalled after a full package update of a Git for Windows SDK).

This commit reflects those changes.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>