Add files reflecting a conflict when updating Pacman packages
When updating a Pacman package and a file has been modified that is
marked "precious" by the Pacman package, Pacman won't overwrite the
file, but instead write the new contents into a file with the `.pacnew`
suffix.

This happens for Git for Windows' SDK because it needs to rewrite the
contents of /etc/nsswitch.conf and of /etc/pacman.conf to work.

Therefore, every time the filesystem or the pacman package is updated,
these files will be accompanied by those `.pacnew` files.

There is really nothing we can do about it, except record their
presence, which this commit does.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>