Add `rsync`
The primary purpose of this repository is to support Continuous
Integration without peppering MSYS2's Pacman repositories with requests,
by simply updating once every night and then using the result for all
jobs that require a Git for Windows SDK.

One really neat trick in building those jobs is to populate a portable
Git from the current SDK in a quick 'n dirty way:

	(cd / &&
	 rsync -Rau $(ARCH=x86_64 BITNESS=64 \
		usr/src/build-extra/make-file-list.sh) \
		/where/ever)

Subsequent build steps may then use that portable Git to run
build-extra's please.sh script e.g. to replace the msys2-runtime in the
SDK (which would not be possible if the script was using that SDK's
msys-2.0.dll...).

This trick requires `rsync`, so let's install it by default.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>