Downgrade mingw-w64-gcc to v7.3.0-2
The upgrade to v8.2.0 was most likely a bit premature. For one, many
function pointer casts in Git for Windows now produce errors. Even
worse, -fstack-protector-strong is broken: apparently, it tries to use
native TLS but that is not even supported by mingw-w64.

So let's roll back that upgrade in one big hurry.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>