Add a sparse-checkout definition for a minimal build environment
Git for Windows' SDK is used for many things, but its primary purpose is
undoubtedly to build and test Git. This task is performed most
prominently in the continuous testing performed in Azure Pipelines.

To help with the bandwidth and time spent by said Pipeline to initialize
a Git for Windows SDK, let's add a sparse-checkout definition that
serves a tiny subset of files, just enough to build and test Git.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>