libxml2: downgrade from 2.9.9-1 to 2.9.8-1
Version 2.9.9-1 is broken in some bad way, making xsltproc.exe fail to
even start because at least the symbols `xmlParserVersion` and
`xmlGenericError` cannot be found, which is somewhat worrisome as such
central functions should not simply just go away.

To work around this, simply downgrade to the last known good version.

Also change `pacman.conf` to prevent this package from being upgraded
for the time being. We can revert this change as soon as
https://github.com/Alexpux/MSYS2-packages/issues/1566 is addressed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>