git-sdk-64-minimal: include the system gitconfig
The main idea of git-sdk-64-minimal is to serve as a very small subset
of the Git for Windows SDK for use in the Azure Pipelines, containing
barely enough to build Git and to run the test suite.

Turns out that we never included the system gitconfig, and relied on the
fact that a full Git for Windows (with `C:\ProgramData\Git\config`) is
installed on those agents. But we no longer read that config, and as a
consequence our minimal SDK's `git.exe` had a different idea of
`core.autocrlf` than the `git.exe` that had checked out the source code.

Let's work around this by including our own system gitconfig.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>