ci-artifacts: also generate & upload the `minimal` artifact
The `git-sdk-64-minmal` artifact is used in _all_ of Git for Windows',
GitGitGadget's and git.git's CI/PR builds. It is therefore of great
importance that we do not upload an artifact that regresses on Git
builds and/or running Git's test suite.

Once all of the CI/PR builds have been ported to GitHub Actions, or at
least changed to no longer access the Azure Pipeline Build Artifact but
instead download the artifact from Azure Blobs (which is a lot faster),
https://dev.azure.com/git-for-windows/git/_build/index?definitionId=22
can be retired in favor of this GitHub Action.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>