ci-artifacts: work around t7300.51 failure
The absolute path of the test's worktree is _just_ long enough for
`FindFirstFileW()` to fail with `ERROR_PATH_NOT_FOUND` (which never
translates into `ENAMETOOLONG`).

Which means that we do not get the error message expected by the test,
but instead "path not found".

Let's use a shorter one by extracting the `git` worktree next to the
top-level directory rather than inside it. That way, it is not that
`FindFirstFileW()` that produces the error, and the test will see the
error message it expects.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>