ci-artifacts: skip builds in forks
It does not make much sense to spend a lot of time building these
artifacts whenever somebody forks git-sdk-64, only to fail in the
uploading step due to a missing secret.

Ideally, we would have an `if: secrets.XYZ` but that does not exist:
https://github.community/t5/GitHub-Actions/If-expression-with-context-variable/td-p/34556

So we do the next-best thing and build only when we're not in a fork.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>