ci-artifacts: let it work in forks, too
So far, it was rather hard-coded to git-for-windows/git-sdk-64.

Even if the current version does not trigger in forks by default, it can
be useful to test things in a fork before opening a PR.

So let's make it easier to do exactly that: now, all contributors need
to do for the workflow to run upon pushing is to remove those
`if: github.event.repository.fork == false` lines.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>