full-sdk: do not use partial clone
The way the `full-sdk` ci-artifact is built does not use a temporary
sparse checkout, but instead uses `git archive` directly on the bare
repository. If said bare repository is a partial clone, each missing
blob is fetched individually (this is a shortcoming of the current
implementation of `git archive`...).

To side-step that, let's just not use a partial clone for the
`full-sdk` ci-artifact.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>