sync: avoid Cygwin heap problem
When running a Cygwin program from another Cygwin program that uses a
different Cygwin runtime version, the dreaded Cygwin heap problem occurs
and things fall apart. Same thing with the MSYS2 runtime.

Now, when we launch Git for Windows' SDK's Bash, it uses one MSYS2
runtime, and if we have an incorrectly-configured `PATH`, we might
execute Git for Windows' (rather than its SDK's) version of, say,
`cygpath.exe`. Which would then try to use a potentially incompatible
MSY2 runtime.

Let's avoid the trouble by configuring the `PATH` correctly, simply by
executing the Bash as a login shell (which forces it to load
`/etc/profile`, which sets the `PATH` as desired).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>