ci: only build the `minimal-sdk` artifacts
The intended use case for the other artifacts was to change the
`setup-git-for-windows-sdk` Action to download these GitHub workflow
artifacts instead of the ones from the Azure Pipeline over at
https://dev.azure.com/git-for-windows/git/_build?definitionId=29&_a=summary.

However, as of v1.3.0, that Action no longer uses build artifacts at all
anymore (except in the `minimal-sdk` case), but instead does a
combination of a partial clone and a parallel checkout for _even faster_
operation.

Therefore, we truly do not need to build the `full-sdk`,
`build-installers` and `makepkg-git` artifacts anymore.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>