ci-artifacts: stop uploading the `minimal-sdk` artifact
Some time ago, this Git for Windows maintainer had the loose idea of
switching the `git-sdk-64-minimal` build from an Azure Pipelines to a
GitHub workflow.

This plan was foiled by one simple fact: while build artifacts can be
downloaded anonymously from Azure Pipelines runs, the same is not true
for GitHub workflow runs.

To remedy that, the idea was explored to upload the artifact to Azure
Blobs instead, but that idea was stopped rather quickly because there
were _so many_ CI builds that the cost was prohibitive.

But that idea was not _really_ stopped altogether, as the `ci-artifacts`
workflow still uploads the artifact to Azure Blobs. Let's drop that,
too, it's of no use.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>