ci-artifacts: run more validations of `minimal-sdk`
The `minimal-sdk` artifact is used in the
`git-for-windows/setup-git-for-windows-sdk` GitHub Action.

This Action is used in Git's own CI builds as well as in Git for
Windows', but also in other workflows, e.g. in
https://github.com/rainers/cv2pdb/actions/workflows/build-and-test.yml

To keep basic functionality working that is assumed to be guaranteed in
the meantime, such as the fact that it comes with a functional GCC that
can build DLL files, which has become a promise, really, let's
specifically test for that.

This is something that the Azure Pipeline already did, and now the
`ci-artifacts` workflow reflects that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>