makepkg-git-i686: include libisl
As of mingw-w64-i686-gcc 12.1.0-3, `cc1.exe` depends on `libisl-23.dll`.

In 44042507c40 (minimal-sdk: include libisl, 2022-07-06),  we already
included this in the `minimal-sdk` build, to fix the regular CI/PR
builds of Git and Git for Windows, and now we also include it in the
`makepkg-git-i686` sparse checkout, to fix building 32-bit Git for
Windows.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>