ci: continuously build Git for Windows' artifacts
This hopefully will prevent last-minute surprises on "Release Day"
(which thankfully haven't been all that common as of late).

This is a heavily edited copy of
https://github.com/git-for-windows/git/blob/v2.37.2.windows.2/.github/workflows/git-artifacts.yml

It would have been nicer to be able to reuse parts of that workflow, of
course, but the fact that we must not use `setup-git-for-windows-sdk` in
this workflow (for obvious reasons) makes this not worth the effort.

To simplify the workflow definition, all GPG and code-signing parts have
been dropped: we are unlikely to ever want to advertise these artifacts
as anything remotely official.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>