git-artifacts: retain artifacts only for a very short period of time
We do not want them to be stored for a long time at all. They are not
artifacts we want to be user-facing, they're more to verify that we can
still release Git for Windows on any given day.

The retention period for the `build-installers` artifact is kept extra
short because it is a huge one (currently ~740MB) and its purpose is to
be used only within the same workflow run (it is reused in the matrix
jobs). It is not intended to ever be downloaded by any user.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>