Fix `git.exe`
A recent update to `libssp-0.dll` now causes a 0xc0000142 when loading
the current `git.exe`. *sigh*

Since that DLL should not be necessary at all anymore, anyway, because
newer `gcc.exe` will not link to it, let's just bite the bullet and
revert to the working version of `libssp-0.dll` for now.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>