ci: verify that the Git and Scalar executables work in Nano Server
It was reported in https://github.com/git-for-windows/git/issues/4052
that Git for Windows v2.38.0 introduced a regression where `git.exe`
could no longer be run in Nano Server containers.

The issue was fixed in https://github.com/git-for-windows/git/pull/4074

Let's add a workflow that verifies that Git for Windows does not regress
on this in the future.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>