ci-artifacts: avoid using PowerShell (or an unintended Bash)
The Git for Windows SDK Bash does the job and the code is easier to
read/write.

We also restrict the `PATH` to avoid picking up e.g. Strawberry Perl or
MSYS2's `gcc`, which are installed on the GitHub build agents, but which
we do not want to use in this workflow.

In the case of the `assorted-validations` job, this quite likely makes a
difference, as we previously ran the test using the Bash from
`C:\Program Files\Git\bin\bash.exe` instead of the `minimal-sdk`'s...

Note: We do need the `GITHUB_PATH` construct to ensure that the correct
Bash is used in `shell: bash` steps because `env: PATH: ...` is
apparently not heeded there.

Avoiding the usage of PowerShell also fixes a problem that started
breaking the `ci-artifacts` builds as of the seemingly unrelated commit
816db2fbe67944d290ff5ea7280ac3ddcb88cdfd. The symptom looks like this:

	▼ Run & .\minimal-sdk\usr\bin\bash.exe -lc @"
	  & .\minimal-sdk\usr\bin\bash.exe -lc @"
	    cd ../git/t &&
	    make T=\"`$(ls -S t[0-9]*.sh | awk \"!((NR+3)%17)\" | tr '\n' \\ )\" prove || {
	      for d in trash*
	      do
		t=`${d#trash directory.}
		echo ===========================
		echo Failed: `$t.sh
		cat test-results/`$t.out
	      done
	      exit 1
	    }
	  "@
	  shell: C:\Program Files\PowerShell\7\pwsh.EXE -command ". '{0}'"
	  env:
	    LC_CTYPE: C.UTF-8
	    GIT_TEST_OPTS: --verbose-log -x --no-chain-lint
	    GIT_PROVE_OPTS: --timer --jobs 8
	    NO_SVN_TESTS: 1
	/usr/bin/bash: -c: line 2: syntax error near unexpected token `('
	/usr/bin/bash: -c: line 2: `  make T=\"$(ls -S t[0-9]*.sh | awk \"!((NR+3)%17)\" | tr '\n' \ )\" prove || {'
	Error: Process completed with exit code 1.

For the full logs, see
https://github.com/git-for-windows/git-sdk-64/actions/runs/3471184593/jobs/5800535064#step:6:22

The most likely explanation is the upgrade from PowerShell v7.2.7 to
v7.3.0: Those are the versions used by the images of the last successful
and of the first failing run. Here are the relevant links:
https://github.com/git-for-windows/git-sdk-64/actions/runs/3470817000/jobs/5799688834#step:1:10
https://github.com/git-for-windows/git-sdk-64/actions/runs/3471184593/jobs/5800535064#step:1:10

Sadly, there is no smoking gun in the release notes of PowerShell
v7.3.0: https://github.com/PowerShell/PowerShell/releases/tag/v7.3.0

By not relying on PowerShell and its quoting/escaping idiosyncracies (it
is easy to lose track of what is needed where, e.g. backticks for dollar
signs, backslashes for double quotes and backslashes), we sidestep this
problem altogether.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>