ci-artifacts: avoid costly compression of the git artifacts
It takes a long time, and the space/bandwidth benefits are not worth it.

In a non-scientific, yet highly indicative test, compressing the
`git-artifacts.tar` file using `xz` took 1m7s and uploading the result
took 2s, while using `gzip` instead took 7s and 20s for compressing and
uploading, respectively. The accumulated timings are therefore ~70s vs
~30s, in favor of using `gzip` compression.

Note: The price is a larger artifact (40.5MB with `gzip` compression, as
compared to 15.6MB with `xz` compression), but since the download time
is 1s, it seems prudent to go with `gzip` compression, still.

Further note that the `upload-artifact` Action implicitly uses `gzip`
compression, but not exclusively, and it also filters out certain
files/directories (e.g. `.git`), therefore we do not want to rely on
that compression. Besides, one other trick up `upload-artifact`'s sleeve
is to do some deduplication, which takes quite a bit of time, time we do
not want to spend here.

Also note: We still use `xz` to compress the `minimal-sdk` artifact. In
the same test as mentioned above, compressing it took 2m6s and uploading
took 8s, while downloading and uncompressing took 5s, compared to 4m5s
of uploading it "uncompressed" (i.e. using the `gzip` compression
implicit in the `upload-artifact` Action) and 1m3s to download it. The
accumulated time is ~135s vs ~240s for compressing+uploading (in favor
of using `xz`) and 5s vs ~60s for downloading+extracting (_also_ in
favor of using `xz`). Combined with the space savings (38.1MB vs 217MB,
_also_ in favor of using `xz`), we will keep using `xz` compression for
the `minimal-sdk` artifact.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>