ci-artifacts: verify that we're using the minimal SDK
It would appear that we might have been testing the wrong thing all the
time because `shell: bash` apparently does not heed `env: PATH` of the
same step.

Unfortunately, there are other build tools installed on GitHub's build
agents, some of which look eerily like Git for Windows' SDK. Therefore,
we can build using `gcc` without ever noticing that it is not ours. Or
for that matter, the Bash might not be ours, either.

Let's make sure that we're testing the minimal SDK and no other tools.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>