minimal-sdk: include `stat.exe`, to support the Coverity Pipeline
At https://dev.azure.com/Git-for-Windows/git/_build?definitionId=35, Git
for Windows maintains automation that submits builds to Coverity for
security scanning.

We just switched that Pipeline to use a minimal SDK (because it really
does not need to spend 4 minutes, 500GB worth of downloaded Git objects
and 2GB worth of disk space, just to compile Git and pre-analyze it for
Coverity).

Among other problems, this uncovered the fact that the pretty basic
`stat.exe` utility was not included in the minimal SDK.

Now, Git's test suite seems to be fine, probably because 1) it runs via
Bash, which has `stat` as a built-in, or 2) errors might be ignored or
misinterpreted.

But let's just include `stat` in minimal SDK.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>