ci-artifacts: run in Pull Requests
I'm about to submit a new Pull Request to integrate these changes. Let's
get some CI goodness.

This requires a bit of a dance when cloning the bare repository because
we still want it to be a bare, shallow _and_ partial clone. But we
cannot do that without a branch name (which PRs do not have). So we need
to fetch the commit OID directly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>