ci-artifacts: avoid the cost of compressing with `xz`
Compressing the `minimal-sdk` artifact with `xz` takes a looong time,
and we only do it to cut down on the time to upload and download the
artifact.

In a quick 'n dirty (read: non-rigorous) test, I got the following
measurements:

	Compression	Upload		Download	Decompression
xz	102s		20s		26s		6s
gz	7s		54s		7s		5s
none	-		281s		184s		-

Now, since we upload `minimal-sdk` in the first job, and download it in
all of the (usually concurrent) matrix jobs, the wall clock cost is
simply the sum of the times in each row:

xz	154s
gz	73s
none	465s

While this test was not very rigorous and there is obviously some
variability, it is still pretty clear that there is one winner here:
`gzip`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>