git-artifacts: accommodate for dynamically-loaded libcurl
Git for Windows recently switched to loading libcurl dynamically, to
allow for selecting from a variety of side-by-side libcurl variants that
are compiled with different SSL/TLS backends.

The mechanics behind that feature involve looking for a libcurl DLL with
the desired name in the `PATH`.

That now poses a problem with the hack where we force `makepkg-mingw` to
use `/mingw64/bin/git.exe` (instead of `/usr/bin/git.exe`, which we do
not want to require to be present) because that libcurl is not on the
path when building an i686 variant of `mingw-w64-git`: in that case,
only `/mingw32/bin/` is in the `PATH`, but not `/mingw64/bin`.

To accommodate for that, let's just extend that hack: It is simply a
script that is written to `/usr/bin/git` and that hands off to
`/mingw64/bin/git.exe`. With this commit, it does so after adding
`/mingw64/bin` to the `PATH` and everything works as intended again.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>