PR build (run the installer): exit with error if the installer failed
When running a PowerShell script, the exit code will be determined
by the last command that was run.

Unfortunately, in the PR build, we run a couple of commands _after_
running the installer, and since we do not check for the exit code
of the installer process, we miss failures.

This is particularly bad because we missed the fact that we failed to
run the 32-bit installer tests ever since we deprecated installing the
32-bit variant on a 64-bit system (such as the GitHub Actions build
agent where the PR build runs...) in c800a3f60 (installer: loudly warn
when installing a 32-bit Git for Windows, 2023-02-21).

Let's check explictly if the installer failed.

Note: We cannot simply use `$?` but need to look at the `ExitCode`
attribute of the process returned by `Start-Process`, _and_ we need to
pass `-PassThru` for this to work, due to a bug in `Start-Process`. See
https://stackoverflow.com/a/16018287/1860823 for a more complete
explanation.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>