ci-artifacts: move minimal-sdk initialization to `/etc/profile`
Instead of using a GitHub Actions-specific shell script to initialize
the shell variables `MSYSTEM` and `PATH`, we now use the standard
`/etc/profile` file. This file is already written by the `please.sh
create-sdk-artifact minimal-sdk` command.

This is the first step to unlock a more general process where the
`minimal-sdk` artifact is made available for wider use by continuously
releasing ready-to-consume archives.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>