ci-artifacts: continuously release the minimal SDK artifact
Taking a page out of MSYS2's book where they release their `srcinfo`
artifact continuously, by replacing the file in the release at
https://github.com/msys2/MINGW-packages/releases/tag/srcinfo-cache, we
publish the minimal SDK artifact to the `ci-artifacts` GitHub release.

It is slighly tricky because GitHub's REST API does not allow to replace
an existing asset atomically, so we have to delete the old one first.

To keep the window small where the asset is not available, we first
upload the new asset to a temporary name, then delete the old one, and
finally rename the new one to the old name.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>