ci-artifacts: provide zip and SFX variants of the minimal SDK
Since `.tar.gz` has its preferred habitat safely outside the Windows
world, let's also provide a `.zip` version of the artifact, and for
good measure even a self-extracting 7-Zip archive (which can be run
everywhere, without having to rely on PowerShell or a native `tar.exe`).

In my tests, both sizes and extract times vary greatly depending on the
used file format. Here are manual measurements as of time of writing:

        .tar.gz     .zip        .7z.exe
Size    95.7MB      84.2MB      41.6MB
Time    4.1s        4.3s        11.1s

Note that I used the native Windows `tar.exe` that is available in the
`C:\Windows\system32` directory which is a BSD tar and comes
preinstalled with Windows build 17063 and later; It is vastly faster
both in extracting `.zip` and `.tar.gz` files (sadly, it has no support
for `.7z` archives) than the corresponding MSYS utilities `unzip` and
`tar` distributed with Git for Windows.

In any scenario where all three formats can be extracted (with roughly
the same speeds as in my tests), and where Defender does not need to
scan the downloaded `.7z.exe` file first (which added up to 8s in my
testing), it therefore depends on the available bandwidth when fetching
from Azure Blobs (where GitHub release assets seem to be stored
nowadays) which one is preferable. At download speeds above 6MB/s, the
`.zip` archive is the fastest to download and extract, below that, the
self-extracting 7-Zip archive provides the best overall speed. Obviously
this should be measured in each particular use case instead of basing
any decision on above-mentioned numbers.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>