pacman: help along with the pacman repo transition
Git for Windows is in the process of migrating to a better home for the
Pacman repo: https://github.com/git-for-windows/pacman-repo.

As part of the migration, the main database file was renamed, and the
`sync` workflow of the `git-sdk-64` was ill-prepared for that. A fix for
that is ready to be deployed as part of the `mingw-w64-git-extra`
package, but its deployment is blocked by the same issue.

Let's work around that by simply copying the database files
(`git-for-windows.db` -> `git-for-windows-x86_64.db` and friends).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>